+++ example_runner.py
@@ -64,6 +64,7 @@ import example.tile.slot
 import example.tile.star
 import example.tile.truchet_triangle
 import example.tile.truchet_triangle_grid_randomized
+import example.tile.truchet_triangle_grid_example
 import example.tile.truchet_circle
 import example.tile.truchet_circle_grid_randomized
 import example.tile.windmill